fix(ui): keep tables and inbox nav routes in workspace shell
