+++ frontend/src/app/workspaces/[workspaceId]/tables/layout.tsx
@@ -1,5 +1,4 @@
 import type { Metadata } from "next"
-import { TablesRouteLayout } from "@/components/tables/tables-route-layout"
 
 export const metadata: Metadata = {
   title: "Tables",
@@ -10,5 +9,5 @@ export default function TablesLayout({
 }: {
   children: React.ReactNode
 }) {
-  return <TablesRouteLayout>{children}</TablesRouteLayout>
+  return children
 }
